Update: 인증 추가 및 S3 이미지 삭제하도록 변경

+++ src/reviews/reviewsController.js
@@ -2,21 +2,28 @@ import { DeleteObjectsCommand } from "@aws-sdk/client-s3";
 import baseResponse from "../../config/baseResponeStatus";
 import { response } from "../../config/response";
 import reviewsService from "./reviewsService";
+import reviewsProvider from "./reviewsProvider";
 
 const reviewsController = {
   deleteReview: async (req, res) => {
     try {
       const {
         params: { reviewId },
+        user: { userId },
       } = req;
 
-      const deleteResult = await reviewsService.deleteReview(reviewId);
-
-      if (deleteResult.error) {
+      const review = await reviewsProvider.findReviewById(reviewId);
+      if (!review) {
         return res.status(404).json(response(baseResponse.REVIEW_NOT_FOUND));
       }
 
-      return res.status(200).json(response(baseResponse.SUCCESS, deleteResult.result));
+      if (!review.author === userId) {
+        return res.status(403).json(response(baseResponse.IS_NOT_REVIEW_OWNER));
+      }
+
+      const deleteResult = await reviewsService.deleteReview(reviewId);
+
+      return res.status(200).json(response(baseResponse.SUCCESS, deleteResult));
     } catch (error) {
       console.log(error);
       return res.status(500).json(response(baseResponse.SERVER_ERROR));
